refactor: IndecisionApp component and child components were moved to /components dir
The indecision app that was previously all under App.js has been
refactored to have each component under their own file.
