+++ src/components/Header.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const Header = ({title, subtitle}) => {
+    return (
+        <div>
+            <h1>{title}</h1>
+            {subtitle && <h2>{subtitle}</h2>}
+        </div>
+    );
+}
+
+Header.defaultProps = {
+    title: 'Indecision'
+};
+  
+export default Header
