+++ src/components/Option.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+const Option = ({optionText, handleDeleteOption}) => {
+    return (
+        <div>
+          {optionText}
+          <button 
+            onClick={(e) => {
+             handleDeleteOption(optionText)
+            }}
+          > 
+            Remove
+          </button>
+        </div>
+    );
+}
+
+export default Option
